requirements.txt
alembic==1.0.7
altgraph==0.16.1
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.7.1
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
facebook-sdk==3.1.0
fake-useragent==0.1.11
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-OAuth==0.13
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.17.1
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
httplib2==0.12.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.0
macholib==1.11
Mako==1.0.7
MarkupSafe==1.1.0
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.0.0
parsel==1.5.1
pefile==2018.8.8
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyInstaller==3.4
pyOpenSSL==18.0.0
pypiwin32==223
PySocks==1.6.8
python-dateutil==2.8.0
python-editor==1.0.4
python-twitter==3.5
python-whois==0.7.1
pythonwhois==2.4.3
pywin32==224
pywin32-ctypes==0.2.0
queuelib==1.5.0
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
Scrapy==1.5.1
scrapy-fake-useragent==1.1.0
scrapy-proxies==0.4
service-identity==18.1.0
six==1.12.0
soupsieve==1.7.2
SQLAlchemy==1.2.18
tweepy==3.7.0
Twisted==18.9.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.1.0
virtualenvwrapper-win==1.2.5
w3lib==1.20.0
Werkzeug==0.14.1
WTForms==2.2.1
zope.interface==4.6.0
